Factor out specification runner
